Moves CombatResolver to package combat.resolver
